# domains/core-platform-services/auth-service/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/academy/notifications": {
      "post": {
        "description": "🔒 Admin / Instructor only",
        "operationId": "NotificationController_createNotification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created"
          }
        },
        "summary": "Create a notification",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/academy/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationController_markNotificationRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mark notification as read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/academy/notifications/{id}": {
      "delete": {
        "operationId": "NotificationController_deleteNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a notification",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "Gateway documentation for frontend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Target user ID",
            "example": "uid_123456"
          },
          "message": {
            "type": "string",
            "description": "Notification message",
            "example": "Your assignment has been graded"
          },
          "type": {
            "type": "string",
            "description": "Notification type",
            "example": "ASSIGNMENT"
          }
        },
        "required": [
          "userId",
          "message",
          "type"
        ]
      },
...
    }
  }
}